packet loss ratio is corretly captured

+++ must-hop/src/tasks/lora.rs
@@ -24,7 +24,7 @@ pub async fn lora_task<Node, T, M, const SIZE: usize, const LEN: usize>(
     node: Node,
     channel: channel::Receiver<'static, M, T, 3>,
     source_id: u8,
-    timeout: u8,
+    timeout: u32,
     max_retries: u8,
     mac: impl MacPolicy<Node, SIZE, LEN>,
 ) where
